Add delete recipe feature.

+++ client/src/utils/hooks/delete-hooks.js
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from 'react';
 import HomeBaristaApi from '../api/home-barista-api';
 
-export const useDeleteBean = (beanId) => {
+export const useDelete = (item, id) => {
   const homeBaristaApi = useMemo(() => new HomeBaristaApi(), []);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
@@ -9,23 +9,39 @@ export const useDeleteBean = (beanId) => {
   const [deleteConfirm, setDeleteConfirm] = useState(false);
 
   useEffect(() => {
-    if(deleteConfirm) {
+    if(deleteConfirm && item === 'bean') {
       setLoading(true);
       const deleteData = async () => {
-        const { data, error } = await homeBaristaApi.deleteBean(beanId)
+        const { data, error } = await homeBaristaApi.deleteBean(id)
         if (data) {
           setSuccess('This bean has been deleted')
           setLoading(false);
         } else {
           setError('Something went wrong... please try again later');
           setLoading(false);
-          console.error(`Error deleting bean with ID ${beanId}: ${error}`)
+          console.error(`Error deleting bean with ID ${id}: ${error}`)
         }
       }
-
       deleteData()
     }
-  }, [homeBaristaApi, deleteConfirm, beanId])
+
+    if(deleteConfirm && item === 'recipe') {
+      setLoading(true);
+      const deleteData = async () => {
+        const { data, error } = await homeBaristaApi.deleteRecipe(id)
+        if (data) {
+          setSuccess('This recipe has been deleted')
+          setLoading(false);
+        } else {
+          setError('Something went wrong... please try again later')
+          setLoading(false);
+          console.error(`Error deleting recipe with ID ${id}: ${error}`)
+        }
+      }
+      deleteData();
+    }
+
+  }, [homeBaristaApi, deleteConfirm, id])
 
   return { loading, error, success, setDeleteConfirm }
 }
